@@ -1,7 +1,7 @@
 {
   "name": "ms",
-  "version": "2.0.0",
-  "description": "Tiny milisecond conversion utility",
+  "version": "2.1.2",
+  "description": "Tiny millisecond conversion utility",
   "repository": "zeit/ms",
   "main": "./index",
   "files": [
@@ -28,10 +28,10 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "eslint": "3.19.0",
+    "eslint": "4.12.1",
     "expect.js": "0.3.1",
-    "husky": "0.13.3",
-    "lint-staged": "3.4.1",
-    "mocha": "3.4.1"
+    "husky": "0.14.3",
+    "lint-staged": "5.0.0",
+    "mocha": "4.0.1"
   }
 }
