@@ -1,10 +1,10 @@
-const { Router } = require('express');
+import { Router } from 'express';
 
 const routes = new Router();
 
 routes.get('/', (req, res) => {
 
-    return res.json({ message: 'Hello World' });
+    return res.json({ message: 'Hello World sky 3' });
 })
 
-module.exports = routes;
+export default routes;
